[run-test] Fix --target and --param parsing.
For argparse, the type value of each argument must be a callable. The previous type was simply pointing to the class that implements the callable, and not to an instance of it.

This should recover the previous functionality for the --target and the --param arguments.

Seems that this is already applied in the build-script, so no changes are necessary there.