CI: adjust the temporary drive, unzip path
This uses `T:` for the temporary drive instead of `S:`.  This is needed
to ensure that the builds on some machines function which already have
the `S` drive in use.  Use an absolute path to `unzip` to allow us to
use the `unzip` from the Git installation rather than relying on MinGW
to provide one.