test: loosen the differentiability witness test
There is currently a difference between the tensorflow branch and the
master branch.  On tensorflow, the differentiability support is merged
into the standard library.  This changes the decoration of the witness.
Loosen the test to accept either.

We should change the tensorflow branch to generate the
`_Differentiation` module with the support and then auto-import the
module in the longer term.  This can be gated by the
`-enable-experimental-autodifferentiation` flag to the driver to gain
the same behaviour on both the branches.