IRGen: Add code to compute a recursive type layout - use  it for value witnesses
Value witnesses for frozen generic types with sub fields will
instantiate generic metadata for those sub fields to compute field
offsets. This is a source of slow down.

This patch instead computes type layouts that can be used to compute the
field offsets recursively.

Value witnesses are implemented in terms of those type layouts and will
no longer instantiate generic metadata (except for resilient types) for
fields.

This will allow for replacing outlined value functions by the witness
implementation without to much overhead in a follow up patch.

This behavior is conditionalized on a flag in IRGenOptions that is
turned off for now.