[android][aarch64] XFAIL IRGen/pic.swift test.
The codegen for Android AArch64 differs in order to the expected one
(and the one that seems to happen in Linux AArch64). The SIL and IR are
the same, but the final codegen is different.

More details in https://bugs.swift.org/browse/SR-12194.

This was introduced with the master rebranch of February 10th 2020.