[ownership] Split out the gathering of users of non-guaranteed values into its own helper function.
This enables me to simplify the code by not using the worklist for
non-guaranteed values (where it is not necessary). I think the non-guaranteed
value handling snuck in over time. I don't think in the non-guaranteed case we
/ever/ used the actual worklist functionality.

This also enabled me to safely, algebraically eliminate unnecessary variables
making the code easier to read (since less indirection).