[ownership] Change BorrowScopeIntroducingValue::{visit,gather}LocalScopeEndingUses to use isConsumingUse instead of checking for isa<EndBorrowInst>().
I am going to be adding support for branch insts/SILPhiArgument to be able to
merge borrow scopes. Thus, the end lifetime for a borrow scope introducing
value, may not be an end_borrow. Luckily, in both cases of br and end_borrow we
model the end of lifetime as a consuming use, so checking for a consuming use
will work both before/after the change.