[ownership] When we flag an over consume, improve error msg by dumping consuming user list.
Specifically, there are a few places in the ownership data flow where before
this patch we just dumped the violating user and the block where the overconsume
occurred. Since we only dumped the block rather than consuming user list, it
could require a little bit of time/energy to identify why the block was
considered to already have a consume in it. So at least by dumping the consuming
user list, the compiler writer has more information to reason about this.