[swift-lang] Fix dependency on sourcekitd module
The link dependency is not sufficient, because this code depends on the
sourcekitd header and module map, which are produced during the build
process.  This started failing after edbe22b63c5, where the dependency
on `sourcekitd-test` was removed. That dependency was itself wrong,
but it happened to be a good enough approximation that we never saw
issues.