[Experiment] Nuke special-case UDRE code for enums
The comment for this is stale. The actual thing this seems to be holding
up is that you can have un-dotted references to enum cases in static
context. This code would eagerly form a pattern for such cases. Now,
for patterns without arguments, we will run the normal lookup path and
form a value of the enum case instead.

I doubt this will have much impact on Swift code, but the original issue
mentions Clang so let's see what happens.