@main: Added $main func to @main type.
Previously, the function was being added to the list of top level decls
in the source file as that list was being iterated, leading to iterator
issues.

Here, the function is instead added to the nominal type decl which is
decorated with @main.  Doing so requires calling the $main function with
the metatype for the nominal type.

The workaround for the iterator invalidation issue which had been
applied previously, namely changing the type of the Decls member of
SourceFile from std::vector<Decl *> to SmallVector<Decl *, 16> is
reverted.